Display a notification when file is not found
